feat: Implement all 15 remaining screens with full navigation

+++ app/src/main/java/com/codewithngoc/instagallery/ui/features/homefeed/postmorebottomsheet/PostMoreBottomSheet.kt
@@ -17,7 +17,7 @@ import com.codewithngoc.instagallery.R
 @Composable
 fun PostMoreBottomSheet(
     onDismiss: () -> Unit,
-    postId: Int
+    postId: Long
 ) {
     val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
